refactor: replace vtcode-tui type definitions with re-exports from vtcode-commons
Replace local type definitions in vtcode-tui with pub use re-exports
from vtcode_commons::ui_protocol. This removes ~670 lines of type
definitions that are now canonical in vtcode-commons.

Affected modules:
- core_tui/types/{style,selection,protocol}.rs
- core_tui/session/config.rs (UiMode, LayoutModeOverride, ReasoningDisplayMode)
- core_tui/style.rs (convert_style, theme_from_color_fields)
- core_tui/app/types/{plan,slash}.rs
- options.rs (SessionSurface, KeyboardProtocolSettings)

InlineListSelection::Reasoning now carries a String instead of a typed
ReasoningEffortLevel, decoupling the protocol layer from config types.
