+++ vtcode-tui/src/core_tui/app/types/slash.rs
@@ -1,14 +1,1 @@
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct SlashCommandItem {
-    pub name: String,
-    pub description: String,
-}
-
-impl SlashCommandItem {
-    pub fn new(name: impl Into<String>, description: impl Into<String>) -> Self {
-        Self {
-            name: name.into(),
-            description: description.into(),
-        }
-    }
-}
+pub use vtcode_commons::ui_protocol::SlashCommandItem;
